Add: CompanyHistory 기본 scss 추가, Banner Component 분리
