+++ public/5.js
@@ -37,6 +37,43 @@ var data = {
     head: "Main Business",
     description: "SNS Commerce Solutions, Fashion Design Platform, SI, Agency"
   }],
+  history: [{
+    year: 2020,
+    historyList: [{
+      month: "12",
+      description: "'찜' 국민 여가생활 필수 어플"
+    }, {
+      month: "11",
+      description: "수많은 여가정보 중 내 취향만 골라서 쏙쏙 ONLY FOR YOU!"
+    }, {
+      month: "10",
+      description: "내가 찜 하고 싶은 그것"
+    }, {
+      month: "09",
+      description: "개개인 취향과 성향을 기반으로 나만을 위한 취향저격 콘텐츠 추천"
+    }, {
+      month: "08",
+      description: "나와 친구의 공통 취향"
+    }]
+  }, {
+    year: 2019,
+    historyList: [{
+      month: "12",
+      description: "이 친구랑 나는 뭘 하고 놀면 좋을까? 공통 취향에 기반한 추천까지"
+    }, {
+      month: "11",
+      description: "다른 사람들은 뭐하면서 놀까"
+    }, {
+      month: "10",
+      description: "나랑 비슷한 성향의 다양한 사람들과 공유하고, 소통하고, 영감도 얻고"
+    }, {
+      month: "09",
+      description: "내 친구 소식 무엇?"
+    }, {
+      month: "08",
+      description: "내 친구들이 가고 싶어하거나 다녀온 곳들이 한 눈에"
+    }]
+  }],
   members: [{
     name: "문성수",
     imgSrc: "https://images.unsplash.com/photo-1522075469751-3a6694fb2f61?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=800&q=80",
