+++ public/css/app.css
@@ -11078,15 +11078,24 @@ a.text-dark:focus {
 }
 
 .CompanyInfo .mainBanner {
-  height: 100vh;
+  position: relative;
   display: flex;
   justify-content: center;
   align-items: center;
   margin-bottom: 50px;
-  background: url("https://images.unsplash.com/photo-1580508158643-4bf9f8da03c6?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1650&q=80") no-repeat 50% 50%/cover;
+}
+
+.CompanyInfo .mainBanner img {
+  width: 100%;
+  height: 100vh;
+  transition: 0.25s ease;
 }
 
 .CompanyInfo .mainBanner p {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
   color: #f8f8f8;
   font-size: 48px;
   font-weight: 600;
@@ -11190,40 +11199,52 @@ a.text-dark:focus {
   font-weight: 600;
 }
 
-.CompanyInfo .brandMission .wrapper ul {
+.CompanyInfo .brandMission .wrapper ul li {
   display: flex;
   flex-direction: column;
-}
-
-.CompanyInfo .brandMission .wrapper ul li {
-  margin-bottom: 30px;
+  margin-bottom: 50px;
   text-align: center;
   font-size: 16px;
   vertical-align: baseline;
 }
 
 .CompanyInfo .brandMission .wrapper ul li span {
-  display: block;
   font-size: 20px;
   font-weight: bolder;
+  margin-bottom: 50px;
+  position: relative;
 }
 
 .CompanyInfo .brandMission .wrapper ul li span.point {
-  display: unset;
   font-size: 24px;
   border-bottom: none;
+  margin-bottom: 0;
+}
+
+.CompanyInfo .brandMission .wrapper ul li span.point:after {
+  display: none;
+}
+
+.CompanyInfo .brandMission .wrapper ul li span:after {
+  content: "";
+  position: absolute;
+  left: 50%;
+  bottom: -25px;
+  transform: translateX(-50%);
+  width: 8px;
+  height: 8px;
+  border-radius: 100px;
+  background-color: #e3342f;
 }
 
 .CompanyInfo .companyProfile {
   background-color: #fff;
-  padding: 50px 0 20px;
-  margin-bottom: 100px;
+  padding: 50px 0 0;
 }
 
 .CompanyInfo .companyProfile .wrapper {
   position: relative;
   padding: 150px;
-  margin-bottom: 150px;
   background: url("https://images.unsplash.com/photo-1500333070776-343c73e1eb4a?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1489&q=80") no-repeat 50% 50%/cover;
 }
 
@@ -11266,6 +11287,48 @@ a.text-dark:focus {
   font-size: 12px;
 }
 
+.CompanyInfo .history {
+  margin-bottom: 150px;
+}
+
+.CompanyInfo .history ul {
+  display: grid;
+  grid-template-columns: auto auto;
+  -moz-column-gap: 50px;
+       column-gap: 50px;
+  max-width: 950px;
+}
+
+.CompanyInfo .history ul li {
+  text-align: center;
+}
+
+.CompanyInfo .history ul li span {
+  display: inline-block;
+  font-size: 32px;
+  margin-bottom: 30px;
+  color: #666;
+}
+
+.CompanyInfo .history ul li table {
+  text-align: left;
+}
+
+.CompanyInfo .history ul li table tr {
+  display: inline-block;
+  margin-bottom: 15px;
+}
+
+.CompanyInfo .history ul li table tr td {
+  word-break: keep-all;
+}
+
+.CompanyInfo .history ul li table tr td:nth-child(1) {
+  color: #e3342f;
+  font-weight: bold;
+  padding-right: 15px;
+}
+
 .CompanyInfo .members {
   margin-bottom: 150px;
 }
@@ -11277,23 +11340,26 @@ a.text-dark:focus {
   grid-row-gap: 20px;
 }
 
-.CompanyInfo .members .wrapper ul li p {
+.CompanyInfo .members .wrapper ul li {
   display: flex;
   justify-content: center;
   align-items: center;
-  flex-direction: column;
   width: 160px;
   height: 160px;
-  border-radius: 100%;
   background-color: #dc3838;
-  color: #fff;
-  font-size: 12px;
+  border-radius: 100%;
   box-shadow: 4px 4px 18px 0px rgba(50, 50, 50, 0.125);
   cursor: default;
 }
 
+.CompanyInfo .members .wrapper ul li p {
+  color: #fff;
+  font-size: 12px;
+}
+
 .CompanyInfo .members .wrapper ul li p span {
   display: block;
+  text-align: center;
   font-size: 16px;
   font-weight: 600;
 }
