+++ resources/sass/CompanyInfo.scss
@@ -18,14 +18,22 @@
     }
 
     .mainBanner {
-        height: 100vh;
+        position: relative;
         display: flex;
         justify-content: center;
         align-items: center;
         margin-bottom: 50px;
-        background: url("https://images.unsplash.com/photo-1580508158643-4bf9f8da03c6?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1650&q=80")
-            no-repeat 50% 50% / cover;
+        img {
+            width: 100%;
+            height: 100vh;
+            transition: 0.25s ease;
+        }
+
         p {
+            position: absolute;
+            top: 50%;
+            left: 50%;
+            transform: translate(-50%, -50%);
             color: #f8f8f8;
             font-size: 48px;
             font-weight: 600;
@@ -126,24 +134,40 @@
 
         .wrapper {
             ul {
-                display: flex;
-                flex-direction: column;
-
                 li {
-                    margin-bottom: 30px;
+                    display: flex;
+                    flex-direction: column;
+                    margin-bottom: 50px;
                     text-align: center;
                     font-size: 16px;
                     vertical-align: baseline;
 
                     span {
-                        display: block;
                         font-size: 20px;
                         font-weight: bolder;
+                        margin-bottom: 50px;
+                        position: relative;
 
                         &.point {
-                            display: unset;
                             font-size: 24px;
                             border-bottom: none;
+                            margin-bottom: 0;
+
+                            &:after {
+                                display: none;
+                            }
+                        }
+
+                        &:after {
+                            content: "";
+                            position: absolute;
+                            left: 50%;
+                            bottom: -25px;
+                            transform: translateX(-50%);
+                            width: 8px;
+                            height: 8px;
+                            border-radius: 100px;
+                            background-color: $red;
                         }
                     }
                 }
@@ -153,13 +177,11 @@
 
     .companyProfile {
         background-color: #fff;
-        padding: 50px 0 20px;
-        margin-bottom: 100px;
+        padding: 50px 0 0;
 
         .wrapper {
             position: relative;
             padding: 150px;
-            margin-bottom: 150px;
             background: url("https://images.unsplash.com/photo-1500333070776-343c73e1eb4a?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1489&q=80")
                 no-repeat 50% 50% / cover;
 
@@ -205,8 +227,47 @@
         }
     }
 
+    .history {
+        margin-bottom: 150px;
+        ul {
+            display: grid;
+            grid-template-columns: auto auto;
+            column-gap: 50px;
+            max-width: 950px;
+
+            li {
+                text-align: center;
+
+                span {
+                    display: inline-block;
+                    font-size: 32px;
+                    margin-bottom: 30px;
+                    color: #666;
+                }
+
+                table {
+                    text-align: left;
+
+                    tr {
+                        display: inline-block;
+                        margin-bottom: 15px;
+                        td {
+                            word-break: keep-all;
+                            &:nth-child(1) {
+                                color: $red;
+                                font-weight: bold;
+                                padding-right: 15px;
+                            }
+                        }
+                    }
+                }
+            }
+        }
+    }
+
     .members {
         margin-bottom: 150px;
+
         .wrapper {
             ul {
                 display: grid;
@@ -215,19 +276,19 @@
                 grid-row-gap: 20px;
 
                 li {
+                    @include alignContentsCenter();
+                    width: 160px;
+                    height: 160px;
+                    background-color: #dc3838;
+                    border-radius: 100%;
+                    box-shadow: 4px 4px 18px 0px rgba(50, 50, 50, 0.125);
+                    cursor: default;
                     p {
-                        @include alignContentsCenter();
-                        flex-direction: column;
-                        width: 160px;
-                        height: 160px;
-                        border-radius: 100%;
-                        background-color: #dc3838;
                         color: #fff;
                         font-size: 12px;
-                        box-shadow: 4px 4px 18px 0px rgba(50, 50, 50, 0.125);
-                        cursor: default;
                         span {
                             display: block;
+                            text-align: center;
                             font-size: 16px;
                             font-weight: 600;
                         }
